requirements.txt
tensorflow-gpu==1.12.0
numpy
pyyaml
termcolor
gym==0.17.1
json_tricks
torch
theano
cached_property
pyopengl
joblib
mako
mujoco_py==0.5.7
